% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_faculty_info.R
\name{add_faculty_info}
\alias{add_faculty_info}
\title{Get the info about a faculty object}
\usage{
add_faculty_info(faculty_ref)
}
\arguments{
\item{faculty_ref}{a reference ID for a faculty}
}
\value{
A character
}
\description{
Get the info about a faculty object
}
\examples{
add_faculty_info("84190")
}
