% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ethics_applications.R
\name{download_ethics_applications}
\alias{download_ethics_applications}
\title{Download all ethics applications}
\usage{
download_ethics_applications(
  file,
  ethics_application_ref = Sys.getenv("ETHICS_APPLICATIONS_REF"),
  sample_n = NULL
)
}
\arguments{
\item{file}{output file}

\item{ethics_application_ref}{the ref for ethics application set by the "ETHICS_APPLICATIONS_REF" env var}

\item{sample_n}{the size of the sampled applications (optional)}
}
\value{
a tibble
}
\description{
Download all ethics applications
}
\examples{
tmp <- tempfile()
download_ethics_applications(tmp, sample_n = 10)
}
