% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_applicants.R
\name{get_applicants}
\alias{get_applicants}
\title{Get the vector of applicants for the given ethics application}
\usage{
get_applicants(application_ref)
}
\arguments{
\item{application_ref}{A object ref}
}
\value{
A character vector of refs
}
\description{
Get the vector of applicants for the given ethics application
}
\examples{
application <- get_applicants("8435x")
}
