% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_content}
\alias{get_content}
\title{A helper function to create a list from a response content}
\usage{
get_content(res)
}
\arguments{
\item{res}{An httr response object}
}
\value{
A list
}
\description{
A helper function to create a list from a response content
}
\examples{
get_object_info("8435x") |>
  get_content()
}
