% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linked_objects.R
\name{get_linked_objects}
\alias{get_linked_objects}
\title{Get linked objects}
\usage{
get_linked_objects(
  obj_ref,
  sources = "NONE",
  key = Sys.getenv("HAPLO_API_KEY"),
  base_URL = Sys.getenv("HAPLO_BASE_URL")
)
}
\arguments{
\item{obj_ref}{Reference ID for the object}

\item{sources}{Haplo API source}

\item{key}{Haplo API Key}

\item{base_URL}{The URL of the Haplo app}
}
\value{
A vector of ref IDs linked to the input object
}
\description{
Get linked objects
}
\examples{
get_linked_objects("84190")
}
