% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_info.R
\name{get_object_info}
\alias{get_object_info}
\title{Get the information about a given object with a ref}
\usage{
get_object_info(ref, sources = "ALL", key = Sys.getenv("HAPLO_API_KEY"))
}
\arguments{
\item{ref}{A ref of a stored object}

\item{sources}{Sources parameter}

\item{key}{API Key}
}
\value{
A httr response object
}
\description{
Get the information about a given object with a ref
}
\examples{
get_object_info("84190")
}
