% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_people_by_faculty.R
\name{get_people_by_faculty}
\alias{get_people_by_faculty}
\title{Get people associated with a faculty}
\usage{
get_people_by_faculty(
  faculty_ref,
  sources = "NONE",
  key = Sys.getenv("HAPLO_API_KEY"),
  base_URL = Sys.getenv("HAPLO_BASE_URL")
)
}
\arguments{
\item{faculty_ref}{A ref for a faculty object}

\item{sources}{Sources as per API}

\item{key}{Haplo API Key}

\item{base_URL}{The URL of the Haplo app}
}
\value{
A character vector of refs
}
\description{
Get people associated with a faculty
}
\examples{
get_people_by_faculty("801yv")
}
