% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_applicant.R
\name{pluck_applicant}
\alias{pluck_applicant}
\title{Pluck a vector of researcher refs from an API response}
\usage{
pluck_applicant(res_content)
}
\arguments{
\item{res_content}{A response content object}
}
\value{
A vector
}
\description{
Pluck a vector of researcher refs from an API response
}
\examples{
# Get an application 8435x
application <- get_object_info("8435x") |>
  get_content()
pluck_applicant(application)

}
